Apk file for app SIMSA
SIMSA stands for Stock Inventory Management System Application. Me and my team (Team 67-The Sims) built it together for the competition Intech-21 done by Universiti Malaya as one of the organizer. We submitted this app on 30/09/2022. We built this app by using software such as Figma, Bravo Studio, Airtable, Firebase and a few other 3rd parties. We are thankful for the experience and knowledge that we gather from joining this competition and looking forward to join another competitions that can challenge us and teach us many other knowledge as well.